fix save funciont when signup, change style for pathList and PathDetail
