+++ client/package.json
@@ -10,7 +10,8 @@
     "axios": "^0.15.3",
     "bootstrap": "^4.0.0-alpha.6",
     "classnames": "^2.2.5",
-    "jwt-decode": "^2.1.0",    
+    "jwt-decode": "^2.1.0",
+    "lodash.merge": "^4.6.0",
     "normalizr": "^3.1.0",
     "ramda": "^0.23.0",
     "react": "^15.4.2",
